Enable (native) unit testing.
Uses the native platform, which compiles with the host system's
compilers.  This was tested on MacOS' GNU c++.

Use the skaygin/ArduinoNative libdep for the Arduino platform
library dependencies; it's the most complete native mock
currently available in PlatformIO and covers almost everything
we use.

Reuses the preexisting stm32 clone of Adafruit's LittleFS for our
filesystem dependency, which seemed to be the easiest path
(though it's also the most demanding of the SPI interface.) Uses
private mocks of SPI and Wire.

The initial two sets of unit tests are fairly trivial, covering
our SHA256 library and mesh::Utils::toHex(), just to show that
everything's working.