promicro_washtastic variant proposal
Although the Washtastic is a DIY board, it's a popular one with increasing deployments. A feature of the Washtastic, to increase its dynamic range, is a non-1:1 voltage bridge. This creates a challenging and frustrating scenario for repeater managers looking for accurate voltage readings, especially in dynamic weather situations.

This PR is a very simple proposed changed, creating a promicro variant "promicro_washtastic" — the only change is the ADC_MULTIPLIER to more accurately report voltage for Wasthastic users looking for a flash that "just works".

We recognize and empathize with efforts to decrease "bloat", but believe that Washtastic proliferation has reached a point where a bespoke firmware may be warranted.

Validated on a Wastastic v4. At charge cutoff, the MeshCore iOS app reported 4.06v while a direct multimeter probe of the 18650 taken at the same time reported 4.059v

Thank you for considering this minor but potentially very helpful addition to the supported devices.